Program to display battery percentage warning

+++ ClassWork/day 4 selection_statement_practice/mobile_battery_warning.py
@@ -18,7 +18,7 @@ Connect Charger Immediately
 
 # Taking battery percentage input from the user
 
-battery_percentage = float(input("Enter the Battery Percentage : "))
+battery_percentage = int(input("Enter the Battery Percentage : "))
 
 print("------------------------------------------------------------------------------------------------------")
